+++ HISTORY.md
@@ -3,5 +3,6 @@
 ## Version 0.0.1 - 29 Jul 2026
 1. Added static data loading
 2. Added card display
+3. Add bidding box to question
 
 ## Version 0.0.0 - 29 Jul 2026