+++ HISTORY.md
@@ -4,5 +4,6 @@
 1. Added static data loading
 2. Added card display
 3. Add bidding box to question
+4. Tidy display of correct/incorrect answers
 
 ## Version 0.0.0 - 29 Jul 2026